feat: model for unique error checking function
